Rework view hierarchy for iOS
