+++ src/MayShow.Shared/App.axaml
@@ -85,8 +85,8 @@
         </Style>
     </Application.Styles>
     <Application.DataTemplates>
-        <DataTemplate DataType="{x:Type viewModels:MainViewModel}">
-            <views:MainView/>
+        <DataTemplate DataType="{x:Type viewModels:CreatePDFReportViewModel}">
+            <views:CreatePDFReportView />
         </DataTemplate>
         <DataTemplate DataType="{x:Type viewModels:WarningDeleteItemViewModel}">
             <views:WarningDeleteItem/>
