+++ src/MayShow.Shared/Views/MainView.axaml.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using Avalonia;
 using Avalonia.Controls;
-using Avalonia.Input;
-using Avalonia.Markup.Xaml;
-using MayShow.ViewModels;
 
 namespace MayShow.Views;
 
@@ -11,34 +6,6 @@ public partial class MainView : UserControl
 {
     public MainView()
     {
-        this.InitializeComponent();
-        LogBlock.PropertyChanged += LogBlock_PropertyChanged;
-        FilesGrid.CellEditEnded += FileCellEditEnded;
-    }
-
-    private void LogBlock_PropertyChanged(object? sender, AvaloniaPropertyChangedEventArgs e)
-    {
-        if (e.Property.ToString() == "Text")
-        {
-            LogScrollView.ScrollToEnd();
-        }
-    }
-
-    public void UnfocusTextbox()
-    {
-        var topLevel = TopLevel.GetTopLevel(this);
-        topLevel?.FocusManager?.ClearFocus();
-        if (DataContext is MainViewModel mvm)
-        {
-            mvm?.HasUnsavedWork = true;
-        }
-    }
-
-    private void FileCellEditEnded(object? sender, DataGridCellEditEndedEventArgs args)
-    {
-        if (args.EditAction == DataGridEditAction.Commit && DataContext is MainViewModel mvm)
-        {
-            mvm?.HasUnsavedWork = true;
-        }
+        InitializeComponent();
     }
 }
