+++ src/MayShow.iOS/MayShow.iOS.csproj
@@ -7,10 +7,10 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Avalonia.iOS" />
+    <PackageReference Include="Avalonia.iOS" Version="$(AvaloniaVersion)" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\MayShow.Desktop\MayShow.Desktop.csproj" />
+    <ProjectReference Include="..\MayShow.Shared\MayShow.Shared.csproj" />
   </ItemGroup>
 </Project>
